transaction service/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
psycopg2-binary
httpx
python-dotenv
python-multipart
